Correção do método delete no Planeta Service para checar se o planeta existe na base antes de excluir.

+++ src/main/java/com/b2wdigital/desafio/starwarsapi/service/impl/PlanetaServiceImpl.java
@@ -55,6 +55,8 @@ public class PlanetaServiceImpl implements PlanetaService {
 
 	@Override
 	public void remove(String id) {
-		planetaRepository.deleteById(id);
+		if(show(id) != null) {
+			planetaRepository.deleteById(id);
+		}
 	}
 }
